Add support for fractional font sizes
Word supports fractional font sizes, for example 8.5pt. But PHPWord doesn't.

+++ src/PhpWord/Settings.php
@@ -386,7 +386,7 @@ class Settings
      */
     public static function setDefaultFontSize($value)
     {
-        $value = (int) $value;
+        $value = (float) $value;
         if ($value > 0) {
             self::$defaultFontSize = $value;
